requirements.txt
asgiref==3.9.0
certifi==2025.7.9
charset-normalizer==3.4.2
Django==5.2.4
django-cors-headers==4.7.0
django-filter==25.1
djangorestframework==3.16.0
djangorestframework-filters==0.11.1
dotenv==0.9.9
gunicorn==23.0.0
idna==3.10
packaging==25.0
pillow==11.3.0
psycopg2-binary==2.9.10
python-dotenv==1.1.1
razorpay==1.4.2
requests==2.32.4
sqlparse==0.5.3
tzdata==2025.2
urllib3==2.5.0
whitenoise==6.9.0
